1.添加相册管理 2.添加用户

+++ src/main/java/com/zhxy/photostudio/service/impl/ServicePackageServiceImpl.java
@@ -78,7 +78,12 @@ public class ServicePackageServiceImpl implements ServicePackageService {
     }
 
     @Override
-    public void delete(Integer id) {
-        servicePackageRepository.deleteById(id);
+    public boolean delete(Integer id) {
+        ServicePackage servicePackage = servicePackageRepository.getOne(id);
+        if (servicePackage.getOrders() == null && servicePackage.getOrders().size() == 0) {
+            servicePackageRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
